plugins: Fixed incorrect callback func signature
The return value must be a uintptr (or not return anything).

+++ internal/msplugins/library.go
@@ -93,7 +93,7 @@ func (o *LibraryPluginCtl) PrettyString(indent string) string {
 	return buf.String()
 }
 
-func (o *LibraryPluginCtl) ReadFromAddr(dst unsafe.Pointer, size uint64, srcAddr uintptr) uint8 {
+func (o *LibraryPluginCtl) ReadFromAddr(dst unsafe.Pointer, size uint64, srcAddr uintptr) uintptr {
 	return 1
 
 	// TODO
